feat: create article on prod
