--- src/features/articles/CreateArticleDialog/EditorBlock/EditorTools.tsx
@@ -1,4 +1,0 @@
-import Header from "@editorjs/header";
-export const EDITOR_TOOLS = {
-  header: Header,
-};
