+++ src/features/articles/CreateArticleDialog/index.tsx
@@ -20,58 +20,9 @@ type Props = {
 };
 
 const CreateArticleDialog = ({ onClose }: Props) => {
-  const ejInstance: any = useRef();
-  const [articleData, setArticleData] = useState<any>({});
+  // const [articleData, setArticleData] = useState<any>({});
 
   const [data, setData] = useState<OutputData>();
-
-  async function createMedia() {
-    try {
-      const apiUrl = `${process.env.NEXT_PUBLIC_BASE_API_URL}media/`;
-
-      const requestBody = {
-        category: "IMAGE",
-      };
-
-      const response = await axios.post(apiUrl, requestBody, {
-        headers: {
-          "Content-Type": "application/json",
-          Authorization: `Bearer ${getToken()}`,
-        },
-      });
-
-      return response?.data?.id;
-    } catch (error) {
-      console.error(error);
-    }
-  }
-
-  async function uploadFile(file: any) {
-    try {
-      const mediaId = await createMedia();
-      if (mediaId === "") {
-        console.log("media error");
-        return;
-      }
-      const formData = new FormData();
-      formData.append("file", file);
-
-      const response = await axios.put(
-        `${process.env.NEXT_PUBLIC_BASE_API_URL}media/${mediaId}/upload`,
-        formData,
-        {
-          headers: {
-            "Content-Type": "multipart/form-data",
-            Authorization: `Bearer ${getToken()}`,
-          },
-        }
-      );
-      return response.data;
-    } catch (error) {
-      // Trate os erros caso ocorram
-      console.error(error);
-    }
-  }
 
   // Função para fazer o download da imagem e usar no copy url editorjs
   async function downloadImage(urlData: string) {
@@ -100,81 +51,6 @@ const CreateArticleDialog = ({ onClose }: Props) => {
     }
   }
 
-  // const initEditor = () => {
-  //   const editor = new EditorJS({
-  //     holder: "editorjs",
-  //     onReady: () => {
-  //       ejInstance.current = editor;
-  //     },
-  //     autofocus: true,
-  //     onChange: async () => {
-  //       let content = await editor.save();
-
-  //       console.log(content);
-  //       setArticleData(content);
-  //     },
-  //     // cria objeto POST
-  //     // atualiza PUT
-  //     //
-  //     tools: {
-  //       header: Header,
-  //       image: {
-  //         class: ImageTool,
-  //         config: {
-  //           /**
-  //            * Custom uploader
-  //            */
-  //           uploader: {
-  //             /**x
-  //              * Upload file to the server and return an uploaded image data
-  //              * @param {File} file - file selected from the device or pasted by drag-n-drop
-  //              * @return {Promise.<{success, file: {url}}>}
-  //              */
-  //             uploadByFile(file: string) {
-  //               return uploadFile(file).then((data) => {
-  //                 return {
-  //                   success: 1,
-  //                   file: {
-  //                     url: data.url,
-  //                     // any other image data you want to store, such as width, height, color, extension, etc
-  //                   },
-  //                 };
-  //               });
-  //             },
-
-  //             // /**
-  //             //  * Send URL-string to the server. Backend should load image by this URL and return an uploaded image data
-  //             //  * @param {string} url - pasted image URL
-  //             //  * @return {Promise.<{success, file: {url}}>}
-  //             //  */
-  //             // uploadByUrl(url: string) {
-  //             //   return downloadImage(url).then((image) => {
-  //             //     return {
-  //             //       success: 1,
-  //             //       file: {
-  //             //         url: image,
-  //             //         // any other image data you want to store, such as width, height, color, extension, etc
-  //             //       },
-  //             //     };
-  //             //   });
-  //             // },
-  //           },
-  //         },
-  //       },
-  //     },
-  //   });
-  // };
-
-  // useEffect(() => {
-  //   if (ejInstance.current === null) {
-  //     initEditor();
-  //   }
-  //   return () => {
-  //     ejInstance?.current?.destroy();
-  //     ejInstance.current = null;
-  //   };
-  // }, []);
-
   //   if (isSuccess) {
   //     return (
   //       <Dialog open onOpenChange={onClose}>
@@ -194,13 +70,13 @@ const CreateArticleDialog = ({ onClose }: Props) => {
 
   async function createArticle() {
     try {
-      const hasHeader = articleData?.blocks.find(
+      const hasHeader = data?.blocks.find(
         (block: any) => block.type === "header"
       );
       if (!hasHeader) {
         return;
       }
-      const hasImage = articleData?.blocks.find(
+      const hasImage = data?.blocks.find(
         (block: any) => block.type === "image"
       );
       if (!hasImage) {
@@ -212,7 +88,7 @@ const CreateArticleDialog = ({ onClose }: Props) => {
 
       const requestBody = {
         title,
-        content: JSON.stringify(articleData),
+        content: JSON.stringify(data),
         images: [],
       };
 
@@ -241,7 +117,7 @@ const CreateArticleDialog = ({ onClose }: Props) => {
               holder="editorjs-container"
             />
 
-            <button onClick={createArticle}>vai porra!!!</button>
+            <button onClick={createArticle}>Enviar</button>
           </div>
         </Dialog.Content>
       </Dialog>
